Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float a;
-    scanf("%f",&a);
-    printf("Area: %f",3.14*a*a);
+    double a;
+    scanf("%lf",&a);
+    printf("Area: %.2lf",3.14*a*a);
     return 0;
 }
